[test, utils] Change priority queue to min queue
Had an epiphany about why the countminsketch was broken. Pretty sure
this was the root of the problem. Still need to change a few other
things.

Also added a unit test for the priority queue. I think the remove method
on the priority queue may not be necessaryleaving it for now