[test]Rewrite simple test for distance
I gave up on trying to order the centriods. Since it occured to me that
there might be two different solutions that are both very close to being
correct, but assign many inputs slightly differently.

I'm now testing for distance sum from each vector to its nearest
centriod in kmeans vs RPHash. With random input the kmeans distance
seems to be about 80% +- 5% of the RPHash distance, need more research
to see if this is good.